Reorganize layout and canvas folders (#120)
