+++ src/content-views/constants.js
@@ -5,13 +5,7 @@ export const LABEL_FONT_SIZE = 11;
 export const MARKER_HEIGHT = 20;
 export const MARKER_TICK_HEIGHT = 8;
 export const MARKER_FONT_SIZE = 10;
-export const MARKER_GUTTER_SIZE = 4;
 export const MARKER_TEXT_PADDING = 8;
-export const BAR_HEIGHT = 16;
-export const BAR_HORIZONTAL_SPACING = 1;
-export const BAR_SPACER_SIZE = 6;
-export const MIN_BAR_WIDTH = 1;
-export const SECTION_GUTTER_SIZE = 4;
 export const COLOR_HOVER_DIM_DELTA = 5;
 
 export const INTERVAL_TIMES = [
@@ -29,41 +23,18 @@ export const INTERVAL_TIMES = [
   5000,
 ];
 export const MIN_INTERVAL_SIZE_PX = 70;
-export const MAX_INTERVAL_SIZE_PX = 140;
 
-export const MOVE_WHEEL_DELTA_THRESHOLD = 1;
-export const ZOOM_WHEEL_DELTA_THRESHOLD = 1;
-export const MIN_ZOOM_LEVEL = 0.25;
-export const MAX_ZOOM_LEVEL = 1000;
-
-export const REACT_PRIORITIES = ['unscheduled', 'high', 'normal', 'low'];
-
-export const ROW_CSS_PIXELS_HEIGHT = 16;
-export const TEXT_CSS_PIXELS_OFFSET_START = 3;
-export const TEXT_CSS_PIXELS_OFFSET_TOP = 11;
-export const FONT_SIZE = 10;
-export const BORDER_OPACITY = 0.4;
-
-export const REACT_GUTTER_SIZE = 0; // Increase to add vertical padding to lanes
-export const REACT_EVENT_ROW_PADDING = 4;
-export const REACT_EVENT_SIZE = 6;
-export const REACT_WORK_SIZE = 9;
-export const REACT_EVENT_BORDER_SIZE = 1;
-export const REACT_WORK_BORDER_SIZE = 1;
+export const EVENT_ROW_PADDING = 4;
+export const EVENT_DIAMETER = 6;
+export const REACT_MEASURE_HEIGHT = 9;
+export const BORDER_SIZE = 1;
 
 export const FLAMECHART_FONT_SIZE = 10;
 export const FLAMECHART_FRAME_HEIGHT = 16;
 export const FLAMECHART_TEXT_PADDING = 3;
 
-export const LABEL_FIXED_WIDTH = LABEL_SIZE + REACT_WORK_BORDER_SIZE;
-export const HEADER_HEIGHT_FIXED = MARKER_HEIGHT + REACT_WORK_BORDER_SIZE;
-export const EVENT_ROW_HEIGHT_FIXED =
-  REACT_EVENT_ROW_PADDING + REACT_EVENT_SIZE + REACT_EVENT_ROW_PADDING;
-
 export const COLORS = Object.freeze({
   BACKGROUND: '#ffffff',
-  OTHER_SCRIPT: '#fff791',
-  OTHER_SCRIPT_HOVER: '#ffea00',
   PRIORITY_BACKGROUND: '#ededf0',
   PRIORITY_BORDER: '#d7d7db',
   PRIORITY_LABEL: '#272727',
@@ -91,6 +62,5 @@ export const COLORS = Object.freeze({
   REACT_SUSPEND: '#a6e59f',
   REACT_SUSPEND_HOVER: '#13bc00',
   REACT_WORK_BORDER: '#ffffff',
-  TIME_MARKER_LINE: '#CAD6DE',
   TIME_MARKER_LABEL: '#18212b',
 });
