+++ src/utils/useSmartTooltip.js
