+++ src/view-base/constants.js
@@ -1,0 +1,4 @@
+export const MOVE_WHEEL_DELTA_THRESHOLD = 1;
+export const ZOOM_WHEEL_DELTA_THRESHOLD = 1;
+export const MIN_ZOOM_LEVEL = 0.25;
+export const MAX_ZOOM_LEVEL = 1000;
